correct parsing and encoding flaws
I was able to install wireshark and a bittorrent client
to observe the protocol. In order to do that I had to find
a basic torrent file which is bencoded. Since it is bencoded,
I could read the file, decode it, and then encode it, write to a file.
Upon comparing the original and the new file, there should be no
differences. Thus, I have reasonable faith in the bencoding
serde implementation.

Included in this commit is the torrent original and the output
from running `cargo run` (main; not tests).